+++ go.mod
@@ -5,4 +5,7 @@ go 1.16
 require (
 	github.com/Masterminds/semver v1.5.0
 	github.com/hashicorp/terraform-config-inspect v0.0.0-20210625153042-09f34846faab
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/spf13/cobra v1.2.1
+	github.com/spf13/viper v1.8.1
 )
